backend/requirements.txt
alembic==1.7.7
click==8.1.3
colorama==0.4.4
ecdsa==0.17.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
pyasn1==0.4.8
python-dotenv==0.20.0
python-jose==3.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.36
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
gunicorn==20.1.0